backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.26.0
scipy>=1.11.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
httpx>=0.27.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
